Break single page into components.
