+++ src/components/contact.js
@@ -1,0 +1,39 @@
+import React from 'react'
+
+export default class Contact extends React.Component {
+  render() {
+    return (
+      <article
+        id="contact"
+        className={`${this.props.article === 'contact' ? 'active' : ''} ${
+          this.props.articleTimeout ? 'timeout' : ''
+        }`}
+        style={{ display: 'none' }}
+      >
+        <h2 className="major">Contact</h2>
+        <form method="post" action="#">
+          <div className="field half first">
+            <label htmlFor="name">Name</label>
+            <input type="text" name="name" id="name" />
+          </div>
+          <div className="field half">
+            <label htmlFor="email">Email</label>
+            <input type="text" name="email" id="email" />
+          </div>
+          <div className="field">
+            <label htmlFor="message">Message</label>
+            <textarea name="message" id="message" rows="4" />
+          </div>
+          <ul className="actions">
+            <li>
+              <input type="submit" value="Send Message" className="special" />
+            </li>
+            <li>
+              <input type="reset" value="Reset" />
+            </li>
+          </ul>
+        </form>
+      </article>
+    )
+  }
+}
